cleanups from Andre
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@3075 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/lyxlex.h
@@ -21,7 +21,7 @@ struct keyword_item {
 	///
 	char const * tag;
 	///
-	short code;
+	int code;
 };
 
 /** Generalized simple lexical analizer.
