+++ src/vspace.C
@@ -204,7 +204,7 @@ bool isValidGlueLength (string const & data, LyXGlueLength * result)
 
 	// To make isValidGlueLength recognize negative values as
 	// the first number this little hack is needed:
-	short val_sign = 1; // positive as default
+	int val_sign = 1; // positive as default
 	switch (buffer[0]) {
 	case '-':
 		lyx_advance(buffer, 1);
@@ -270,7 +270,7 @@ bool isValidLength(string const & data, LyXLength * result)
 
 	// To make isValidLength recognize negative values
 	// this little hack is needed:
-	short val_sign = 1; // positive as default
+	int val_sign = 1; // positive as default
 	switch (buffer[0]) {
 	case '-':
 		lyx_advance(buffer, 1);
